fix(flat): Flat[Long].decode truncated values >= 2^34
(w & 0x7f) << shl shifted an Int, so shift counts >= 32 wrapped mod 32
and any Long needing 6+ varint groups decoded to garbage. Widen to Long
before shifting.

+++ scalus-core/shared/src/main/scala/scalus/serialization/flat/package.scala
@@ -119,11 +119,11 @@ package object flat:
             var r = 0L
             var shl = 0
             while (w & 0x80) != 0 do
-                r = r | ((w & 0x7f) << shl)
+                r = r | ((w & 0x7f).toLong << shl)
                 shl += 7
                 w = decode.bits8(8)
 
-            r = r | ((w & 0x7f) << shl)
+            r = r | ((w & 0x7f).toLong << shl)
             zagZig(r)
 
     given Flat[BigInt] with
